Re-write lambda using Flask and apig-wsgi So the main lambda code will be lambda-agnostic, i.e. we will be able to easily run the same code in e.g. Docker with minimum changes
