+++ requirements-deploy.txt
@@ -1,0 +1,2 @@
+Flask~=3.1.0
+apig-wsgi~=2.19.0
