+++ requirements.txt
@@ -2,3 +2,6 @@ aws-cdk-lib==2.164.0
 constructs>=10.0.0,<11.0.0
 
 boto3~=1.35.49
+
+Flask~=3.1.0
+apig-wsgi~=2.19.0
